--- src/utils/datetime.ts
@@ -1,11 +1,0 @@
-import {Calendar, HourCycle} from 'enums/datetime'
-
-export const getCalendars = (): Calendar[] => Object.values(Calendar)
-
-export const getHourCycles = (): Intl.LocaleHourCycleKey[] =>
-  Object.values(HourCycle)
-
-export default {
-  getCalendars,
-  getHourCycles,
-}
