--- src/utils/location.ts
@@ -1,7 +1,0 @@
-import {Region} from 'enums/location'
-
-export const getRegions = (): Region[] => Object.values(Region)
-
-export default {
-  getRegions,
-}
